The "pages" folder has been removed. The "dependencies" components have been nested. CheckBox and Select components have been removed. RatingDiv has been moved to the "Ratings" folder.
