--- src/components/CheckBox.js
@@ -1,18 +1,0 @@
-export function CheckBox({
-    labelName, 
-    paramName, 
-    checked, 
-    onChange
-}) {
-    return (
-        <div>
-            <label htmlFor={paramName}>{labelName}:</label>
-            <input
-                type="checkbox"
-                name={paramName}
-                checked={checked}
-                onChange={onChange}
-            />
-        </div>
-    );
-};
