+++ src/components/HotelDetailsPage/Hotel/Hotel.module.css
